feat: Changes in the main file
Add custom route animation
Divide app class and main function into different files
Move constants
